+++ remove-http.php
@@ -10,7 +10,7 @@
  */
 
 // If accessed directly, exit
-if ( ! defined( 'ABSPATH' ) ) exit;
+if ( !defined( 'ABSPATH' ) ) exit;
 
 class Fact_Maven_Remove_HTTP {
 
@@ -22,7 +22,7 @@ class Fact_Maven_Remove_HTTP {
         // Get plugin's metadata
         $this->plugin = get_plugin_data( __FILE__ );
         // If the plugin version is lower or not defined, remove plugin options
-        if ( ( get_option( 'factmaven_rhttp_version' ) < $this->plugin['Version'] ) || ! get_option( 'factmaven_rhttp_version' ) ) {
+        if ( ( get_option( 'factmaven_rhttp_version' ) < $this->plugin['Version'] ) || !get_option( 'factmaven_rhttp_version' ) ) {
             // Remove options with the prefix "factmaven_rhttp_"
             foreach ( wp_load_alloptions() as $option => $value ) {
                 if ( strpos( $option, 'factmaven_rhttp' ) === 0 ) delete_option( $option );
@@ -35,7 +35,6 @@ class Fact_Maven_Remove_HTTP {
         $this->option = get_option( 'factmaven_rhttp' );
         // Set default options
         if ( empty( $this->option['format'] ) ) $this->option['format'] = 'protocol-relative';
-        if ( empty( $this->option['dashboard'] ) ) $this->option['dashboard'] = '0';
         if ( empty( $this->option['ignore'] ) ) $this->option['ignore'] = '';
         if ( empty( $this->option['external'] ) ) $this->option['external'] = '0';
         // Add link to plugin settings
@@ -44,14 +43,8 @@ class Fact_Maven_Remove_HTTP {
         add_filter( 'admin_init', array( $this, 'settings_field' ), 10, 1 );
         // Relocate settings field using jQuery
         add_action( 'admin_footer', array( $this, 'settings_location' ), 10, 1 );
-        // If 'Apply to Admin Dashboard' is checked, apply to entire website
-        if ( $this->option['dashboard'] != '1' ) {
-            add_action( 'wp_loaded', array( $this, 'protocol_relative' ) , PHP_INT_MAX, 1 );
-        }
-        // Else, remove HTTP and HTTPS protocols on the frontend, ignore admin dashboard
-        else {
-            if ( ! is_admin() && defined( 'DOING_AJAX' ) ) add_action( 'wp_loaded', array( $this, 'protocol_relative' ) , PHP_INT_MAX, 1 );
-        }
+        // Protocol is only removed on frontend, ignore Admin Dashboard
+        if ( !is_admin() && defined( 'DOING_AJAX' ) ) add_action( 'wp_loaded', array( $this, 'protocol_relative' ) , PHP_INT_MAX, 1 );
     }
 
     public function settings_link( $links, $file ) {
@@ -84,7 +77,6 @@ class Fact_Maven_Remove_HTTP {
             <label><input type="radio" name="factmaven_rhttp[format]" value="relative" <?php checked( 'relative', $this->option['format'] ); ?>> <span class="date-time-text">Relative</span><code>/sample-post/</code></label>
             <p class="description" id="format-description">Relative format will only affect internal links.</p>
             <br>
-            <label><input name="factmaven_rhttp[dashboard]" type="checkbox" value="1" <?php checked( '1', $this->option['dashboard'] ); ?>>Apply to Admin Dashboard</label><br>
             <!-- <label>Exclude the following domains and URLs, internal or external. Enter each link on a new line.</label>
             <p><textarea name="factmaven_rhttp[ignore]" rows="5" cols="80" class="code" placeholder="<?php echo preg_replace( '/^https?:\/\//', '', home_url() ); ?>&#x0a;<?php echo home_url(); ?>/sample-post"><?php echo $this->option['ignore']; ?></textarea></p> -->
             <label><input name="factmaven_rhttp[external]" type="checkbox" value="1" <?php checked( '1', $this->option['external'] ); ?>>Ignore all external links</label>